feat : 이메일 존재 여부 확인 로직 추가

+++ src/main/java/masil/backend/modules/member/service/MemberEmailVerificationHighService.java
@@ -2,46 +2,50 @@ package masil.backend.modules.member.service;
 
 import static masil.backend.modules.member.exception.MemberExceptionType.EMAIL_CODE_EXPIRED;
 import static masil.backend.modules.member.exception.MemberExceptionType.EMAIL_CODE_NOT_MATCH;
+import static masil.backend.modules.member.exception.MemberExceptionType.INVALID_EMAIL;
 
 import java.security.SecureRandom;
 import java.time.LocalDateTime;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import masil.backend.modules.member.entity.MemberEmailVerification;
 import masil.backend.modules.member.exception.MemberException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@Slf4j
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class MemberEmailVerificationHighService {
     private final MemberEmailVerificationLowService memberEmailVerificationLowService;
     private final MemberLowService memberLowService;
+    private final EmailValidationService emailValidationService;
 
     private static final SecureRandom random = new SecureRandom();
-
 
     public void sendVerificationCode(final String email) {
         // 1. 이미 가입된 이메일인지 확인
         memberLowService.checkIsDuplicateEmail(email);
 
-        // 2. 기존 인증 코드가 있으면 삭제
+        // 2. 실제 존재하는 이메일인지 확인
+        if (!emailValidationService.isEmailValid(email)) {
+            throw new MemberException(INVALID_EMAIL);
+        }
+
+        // 3. 기존 인증 코드가 있으면 삭제
         if (memberEmailVerificationLowService.existsEmailVerification(email)) {
             memberEmailVerificationLowService.deleteEmailVerification(email);
         }
 
-        // 3. 6자리 랜덤 코드 생성
+        // 4. 6자리 랜덤 코드 생성
         final String code = generateCode();
 
-        // 4. 만료 시간 설정 (5분 후)
+        // 5. 만료 시간 설정 (5분 후)
         final LocalDateTime expiresAt = LocalDateTime.now().plusMinutes(5);
 
-        // 5. DB에 저장
+        // 6. DB에 저장
         memberEmailVerificationLowService.saveEmailVerification(email, code, expiresAt);
 
-        // 6. 이메일 발송
+        // 7. 이메일 발송
         memberEmailVerificationLowService.sendEmail(email, code);
     }
